+++ app/src/main/java/ru/ssau/population/model/ChartState.kt
@@ -1,6 +1,6 @@
 package ru.ssau.population.model
 
 data class ChartState(
-    val t: List<Double>,        // моменты времени
-    val y: List<List<Double>>,  // список кривых (у каждой кривой список точек)
+    val t: List<Float>,        // моменты времени
+    val y: List<List<Float>>,  // список кривых (у каждой кривой список точек)
 )
