+++ app/src/main/java/ru/ssau/population/model/PopulationState.kt
@@ -3,5 +3,5 @@ package ru.ssau.population.model
 // одна популяция в конкретный момент времени
 interface PopulationState {
     val population: PopulationStats
-    val count: Double
+    val count: Float
 }
